+++ projects/ui/src/lib/components/icon/index.ts
@@ -1,4 +1,5 @@
 export * from './kui-icon.component';
+export * from './kui-icon-lucide-default';
 export * from './kui-icon-registry.token';
 export * from './kui-icon-source.type';
 export * from './provide-kui-icons';